fix gptq unit test

+++ test/srt/test_gptqmodel_dynamic.py
@@ -106,7 +106,7 @@ class TestGPTQModelDynamic(CustomTestCase):
             cls.model,
             cls.base_url,
             timeout=DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
-            other_args=["--dtype", "float16"],
+            other_args=["--dtype", "bfloat16"],
         )
 
     @classmethod
